+++ DrillingRig.Commands/AinSettings/IAinSettings.cs
@@ -12,8 +12,8 @@ namespace DrillingRig.Commands.AinSettings {
 		short Imax { get; }
 		short UdcMax { get; }
 		short UdcMin { get; }
-		short Fnom { get; }
-		short Fmax { get; }
+		double Fnom { get; }
+		double Fmax { get; }
 
 		short DflLim { get; }
 		short FlMinMin { get; }
@@ -50,7 +50,8 @@ namespace DrillingRig.Commands.AinSettings {
 		short AccDfDt { get; }
 		short DecDfDt { get; }
 
-		short Unom { get; }
+		//short Unom { get; }
+		double Unom { get; }
 
 		/// <summary>
 		/// Постоянная времени регулятора компенсации потока
@@ -58,7 +59,7 @@ namespace DrillingRig.Commands.AinSettings {
 		short TauFlLim { get; }
 
 		short Rs { get; }
-		short Fmin { get; }
+		double Fmin { get; }
 
 		short TauM { get; }
 		short TauF { get; }
